requirements.txt
Pillow
opencv-python
numpy
os